+++ lib/Query.php
@@ -5,9 +5,9 @@ namespace SearchEngine;
 /**
  * SearchEngine Query class
  *
- * @version 0.3.0
+ * @version 0.3.1
  * @author Teppo Koivula <teppo.koivula@gmail.com>
- * @license Mozilla Public License v2.0 http://mozilla.org/MPL/2.0/
+ * @license Mozilla Public License v2.0 https://mozilla.org/MPL/2.0/
  *
  * @property-read string $selector Final selector string.
  * @property-read string $resultsString Results rendered as a list (PageArray).
@@ -169,10 +169,10 @@ class Query extends Base {
                 return !empty($this->results) && method_exists($this->results, '___getMarkup') ? $this->results->render() : '';
                 break;
             case 'resultsCount':
-                return !empty($this->results) ? $this->results->count() : '';
+                return !empty($this->results) ? $this->results->count() : 0;
                 break;
             case 'resultsTotal':
-                return !empty($this->results) ? $this->results->getTotal() : '';
+                return !empty($this->results) ? $this->results->getTotal() : 0;
                 break;
             case 'pager':
             case 'resultsPager':