+++ lib/Renderer.php
@@ -17,7 +17,7 @@ use ProcessWire\WireException;
  *
  * @version 0.5.1
  * @author Teppo Koivula <teppo.koivula@gmail.com>
- * @license Mozilla Public License v2.0 http://mozilla.org/MPL/2.0/
+ * @license Mozilla Public License v2.0 https://mozilla.org/MPL/2.0/
  */
 class Renderer extends Base {